Improving tests
