+++ test/db-migrations/20250415-01-client-audit-remainder.spec.js
@@ -1,20 +1,12 @@
-const { hash, randomBytes } = require('node:crypto');
-
 const { // eslint-disable-line object-curly-newline
   assertTableSchema,
   assertTableContents,
   describeMigration,
 } = require('./utils'); // eslint-disable-line object-curly-newline
 
-describeMigration('20250415-01-client-audit-remainder', ({ runMigrationBeingTested }) => {
-  const aBlobWith = props => {
-    const randomContent = randomBytes(100);
-    const md5 = hash('md5',  randomContent); // eslint-disable-line no-multi-spaces
-    const sha = hash('sha1', randomContent);
-    return { md5, sha, ...props };
-  };
-  const aBlob = () => aBlobWith({});
+const { aBlob } = require('./fixtures');
 
+describeMigration('20250415-01-client-audit-remainder', ({ runMigrationBeingTested }) => {
   before(async () => {
     await assertTableSchema('client_audits',
       { column_name: 'blobId',        is_nullable: 'NO', data_type: 'integer' }, // eslint-disable-line no-multi-spaces
@@ -38,7 +30,7 @@ describeMigration('20250415-01-client-audit-remainder', ({ runMigrationBeingTest
       RETURNING id
     `);
 
-    await db.oneFirst(sql`
+    await db.any(sql`
       INSERT INTO client_audits (
       "blobId",
       remainder
@@ -47,13 +39,12 @@ describeMigration('20250415-01-client-audit-remainder', ({ runMigrationBeingTest
       ${blobId},
       '{"user": "test-user", "change-reason": "test-reason", "unknown-value": "test-unknown"}'::jsonb
       )
-      returning "blobId"
     `);
 
     await runMigrationBeingTested();
   });
 
-  it('should add columns to client audit table', async () => {
+  it.only('should add columns to client audit table', async () => {
     await assertTableSchema('client_audits',
       { column_name: 'blobId',        is_nullable: 'NO', data_type: 'integer' }, // eslint-disable-line no-multi-spaces
       { column_name: 'event',         is_nullable: 'YES', data_type: 'text' },   // eslint-disable-line no-multi-spaces
@@ -71,7 +62,7 @@ describeMigration('20250415-01-client-audit-remainder', ({ runMigrationBeingTest
     );
   });
 
-  it('should move value from remainder to new columns', async () => {
+  it.only('should move value from remainder to new columns', async () => {
     await assertTableContents('client_audits', {
       'user': 'test-user', // eslint-disable-line quote-props
       'change-reason': 'test-reason',
